refactor(react/backup): split backup form into seperated forms component
