+++ app/react/portainer/settings/SettingsView/BackupSettingsView/BackupFormS3.validation.ts
@@ -2,11 +2,6 @@ import { object, string, boolean } from 'yup';
 
 export function validationSchema() {
   return object().shape({
-    passwordProtect: boolean(),
-    password: string().when('passwordProtect', {
-      is: true,
-      then: (schema) => schema.required('This field is required.'),
-    }),
     passwordProtectS3: boolean(),
     passwordS3: string().when('passwordProtectS3', {
       is: true,
